Create BTreeExercises.h
The header file for some extra Binary Search Tree functions and their implementations. Functions include checking whether the tree is complete (this means that every subtree has a left and right child or that the tree has a left child as far as possible and no right children), mirroring the existing binary search tree, counting the number of nodes.